// src/lib/token-types.ts
/* eslint-disable camelcase */
export enum TOKEN_TYPES {
  variable = 'variable',
  variable_entity = 'variable_entity',
  variable_language = 'variable_language',
  constant = 'constant',
  constant_entity = 'constant_entity',
  constant_language = 'constant_language',
  function = 'function',
  function_entity = 'function_entity',
  function_parameters = 'function_parameters',
  class_entity = 'class_entity',
  class_inherited = 'class_inherited',
  class_members = 'class_members',
  annotation = 'annotation',
  comment = 'comment',
  numeric = 'numeric',
  string = 'string',
  string_character_escape = 'string_character_escape',
  string_placeholder = 'string_placeholder',
  string_quotes = 'string_quotes',
  string_regexp = 'string_regexp',
  string_multiline = 'string_multiline',
  string_continuation = 'string_continuation',
  string_interpolation = 'string_interpolation',
  punctuation = 'punctuation',
  separator = 'separator',
  accessor = 'accessor',
  terminator = 'terminator',
  brackets = 'brackets',
  operator = 'operator',
  operator_word = 'operator_word',
  keyword = 'keyword',
  keyword_symbol = 'keyword_symbol',
  keyword_control = 'keyword_control',
  keyword_flow = 'keyword_flow',
  keyword_import = 'keyword_import',
  keyword_declaration = 'keyword_declaration',
  storage_data_type = 'storage_data_type',
  storage_modifier = 'storage_modifier',
  support = 'support',
  support_constant = 'support_constant',
  support_function = 'support_function',
  support_data_type = 'support_data_type',
  support_class = 'support_class',
  xml_text = 'xml_text',
  xml_tag = 'xml_tag',
  xml_attribute = 'xml_attribute',
  markup = 'markup',
  markup_heading = 'markup_heading',
  markup_list = 'markup_list',
  markup_bold = 'markup_bold',
  markup_italic = 'markup_italic',
  markup_underline = 'markup_underline',
  markup_link = 'markup_link',
  markup_inserted = 'markup_inserted',
  markup_deleted = 'markup_deleted',
  markup_quote = 'markup_quote',
  markup_raw = 'markup_raw',
  markup_info = 'markup_info',
  markup_warning = 'markup_warning',
  markup_error = 'markup_error',
  invalid = 'invalid'
}